Added functions needed to access the information about the actual number of transferred bytes. Unfortunately, it seems that the dmaengine does not offer the necessary functionality.
There was a discussion about adding the necessary field "transferred" to the dma_async_tx_descriptor:
http://www.spinics.net/lists/linux-usb/msg89942.html
However the proposal was rejected, and using the dmaengine_tx_state followed by reading
the "residue" field from the "dma_tx_state" was suggested:
http://www.spinics.net/lists/linux-usb/msg90285.html
It seems however that the current kernel does not support that functionality.
The "residue" field in the "dma_tx_state" is set to 0 after the transfer is completed, regardless of the
actual number of transferred bytes.

The problem was reported to Xilinx forum: https://forums.xilinx.com/t5/Embedded-Linux/AXI-DMA-checking-the-number-of-really-transferred-bytes-for/m-p/715662#M16802